create a cli module
