+++ main.go
@@ -2,7 +2,6 @@ package main
 
 import (
 	"context"
-	"flag"
 	"io/ioutil"
 	"log"
 	"os"
@@ -88,37 +87,6 @@ func main() {
 		go startContainers(ctx, k, v, networkID, networkName, &wg)
 	}
 	wg.Wait()
-}
-
-func Cli() {
-	var showVersion bool
-	var up bool
-
-	flag.BoolVar(
-		&showVersion,
-		"version",
-		false,
-		"Show version information.")
-	flag.BoolVar(
-		&showVersion,
-		"v",
-		false,
-		"Show version information.")
-	flag.BoolVar(
-		&up,
-		"up",
-		false,
-		"Builds, re/creates, starts, and attaches to containers for a service")
-
-	flag.Parse()
-
-	if showVersion {
-		log.Println("Meli version:", Version)
-		os.Exit(0)
-	}
-	if !up {
-		os.Exit(0)
-	}
 }
 
 func fakestartContainers(ctx context.Context, k string, s serviceConfig, networkName, networkID string, wg *sync.WaitGroup) {
